backend/requirements.txt
schedule
pyautogui
pygetwindow
pytesseract
Pillow
